Change product.unit_price data type from INTEGER to REAL.

+++ sql/LtwDb.sql
@@ -45,7 +45,7 @@ CREATE TABLE IF NOT EXISTS customer (
 CREATE TABLE IF NOT EXISTS product (
     id INTEGER PRIMARY KEY AUTOINCREMENT,
     description CHAR(50) NOT NULL,
-    unit_price INTEGER
+    unit_price REAL
 );
 
 CREATE TABLE IF NOT EXISTS invoice (
